src/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lerobot[all]>=0.3.2",
    "pillow>=12.0.0",
    "pyinstaller>=6.17.0",
    "pymycobot>=4.0.3",
    "torch>=2.9.1",
    "xarm>=0.0.4",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.17.0",
]
